feat(flow): support node and edge highlight when hover(#OSP-572)

+++ packages/flow/src/flow-node.component.ts
@@ -1,13 +1,4 @@
-import {
-    ChangeDetectionStrategy,
-    ChangeDetectorRef,
-    Component,
-    ComponentRef,
-    OnDestroy,
-    OnInit,
-    Renderer2,
-    ViewContainerRef
-} from '@angular/core';
+import { ChangeDetectionStrategy, ChangeDetectorRef, Component, OnDestroy, OnInit, Renderer2, ViewContainerRef } from '@angular/core';
 import { TextManage } from '@plait/text';
 import { PlaitPluginElementComponent, PlaitPluginElementContext, PlaitBoard, normalizePoint, createG, OnContextChanged } from '@plait/core';
 import { RoughSVG } from 'roughjs/bin/svg';
@@ -15,6 +6,7 @@ import { drawNodeHandles } from './draw/handle';
 import { drawActiveMask, drawNode } from './draw/node';
 import { FlowNode } from './interfaces/node';
 import { FlowBaseData } from './interfaces/element';
+import { FlowRenderMode } from './public-api';
 
 @Component({
     selector: 'plait-flow-node',
@@ -33,7 +25,7 @@ export class FlowNodeComponent<T extends FlowBaseData = FlowBaseData> extends Pl
 
     handlesG: SVGGElement | null = null;
 
-    elementG!: SVGGElement;
+    activeG: SVGGElement | null = null;
 
     constructor(public cdr: ChangeDetectorRef, public viewContainerRef: ViewContainerRef, public render2: Renderer2) {
         super(cdr);
@@ -48,16 +40,16 @@ export class FlowNodeComponent<T extends FlowBaseData = FlowBaseData> extends Pl
             return { x, y, width, height };
         });
         this.roughSVG = PlaitBoard.getRoughSVG(this.board);
-        this.elementG = createG();
+        this.activeG = createG();
         this.drawElement();
     }
 
     onContextChanged(value: PlaitPluginElementContext<FlowNode, PlaitBoard>, previous: PlaitPluginElementContext<FlowNode, PlaitBoard>) {
         if (value.element !== previous.element && this.initialized) {
-            this.drawElement(value.element, value.selected);
+            this.drawElement(value.element, value.selected ? FlowRenderMode.active : FlowRenderMode.default);
         }
         if (value.selected) {
-            this.drawActiveElement(this.element);
+            this.drawElement(this.element, value.selected ? FlowRenderMode.active : FlowRenderMode.default);
         } else if (this.initialized) {
             if (previous.selected) {
                 this.destroyActiveMask();
@@ -66,26 +58,33 @@ export class FlowNodeComponent<T extends FlowBaseData = FlowBaseData> extends Pl
         }
     }
 
-    drawElement(element: FlowNode = this.element, active = false) {
-        this.updateElement(element, active);
-        this.g.append(this.elementG!);
+    drawElement(element: FlowNode = this.element, mode: FlowRenderMode = FlowRenderMode.default) {
+        this.getElement(element, mode);
+        if (mode === FlowRenderMode.default) {
+            this.g.append(this.nodeG!);
+            this.g.append(this.textManage.g);
+            this.activeG?.remove();
+        } else {
+            this.activeG?.append(this.nodeG!);
+            this.activeG?.append(this.textManage.g);
+            if (mode === FlowRenderMode.active) {
+                this.activeG?.prepend(this.activeMaskG!);
+            }
+            this.activeG?.append(this.handlesG!);
+            PlaitBoard.getElementHostActive(this.board).append(this.activeG!);
+        }
     }
 
-    drawActiveElement(element: FlowNode = this.element, active = true) {
-        this.updateElement(element, active);
-        PlaitBoard.getElementHostActive(this.board).append(this.elementG!);
-    }
-
-    drawElementG(element: FlowNode = this.element) {
+    drawNode(element: FlowNode = this.element) {
         this.destroyElement();
         this.nodeG = drawNode(this.roughSVG, element);
-        this.elementG.append(this.nodeG);
     }
 
-    drawActiveMask(element: FlowNode = this.element) {
+    drawActiveMask(element: FlowNode = this.element, mode: FlowRenderMode = FlowRenderMode.default) {
         this.destroyActiveMask();
-        this.activeMaskG = drawActiveMask(this.roughSVG, element);
-        this.elementG.prepend(this.activeMaskG);
+        if (mode === FlowRenderMode.active) {
+            this.activeMaskG = drawActiveMask(this.roughSVG, element);
+        }
     }
 
     drawRichtext(element: FlowNode = this.element) {
@@ -93,29 +92,27 @@ export class FlowNodeComponent<T extends FlowBaseData = FlowBaseData> extends Pl
         if (element.data?.text) {
             this.textManage.draw(element.data.text);
             this.textManage.g.classList.add('flow-node-richtext');
-            this.elementG.append(this.textManage.g);
         }
     }
 
-    drawHandles(element: FlowNode = this.element) {
-        this.destroyHandles();
-        const handles = drawNodeHandles(this.roughSVG, element);
-        this.handlesG = createG();
-        handles.map(item => {
-            this.handlesG?.append(item);
-            this.render2.addClass(item, 'flow-handle');
-        });
-        this.handlesG?.setAttribute('stroke-linecap', 'round');
-        this.elementG.append(this.handlesG);
+    drawHandles(element: FlowNode = this.element, mode: FlowRenderMode = FlowRenderMode.default) {
+        if (mode !== FlowRenderMode.default) {
+            this.destroyHandles();
+            const handles = drawNodeHandles(this.roughSVG, element);
+            this.handlesG = createG();
+            handles.map(item => {
+                this.handlesG?.append(item);
+                this.render2.addClass(item, 'flow-handle');
+            });
+            this.handlesG?.setAttribute('stroke-linecap', 'round');
+        }
     }
 
-    updateElement(element: FlowNode = this.element, isActive = false) {
-        this.drawElementG(element);
+    getElement(element: FlowNode = this.element, mode: FlowRenderMode = FlowRenderMode.default) {
+        this.drawNode(element);
         this.drawRichtext(element);
-        if (isActive) {
-            this.drawActiveMask(element);
-            this.drawHandles(element);
-        }
+        this.drawActiveMask(element, mode);
+        this.drawHandles(element, mode);
     }
 
     destroyHandles() {
@@ -141,5 +138,11 @@ export class FlowNodeComponent<T extends FlowBaseData = FlowBaseData> extends Pl
 
     destroyRichtext() {
         this.textManage.destroy();
+    }
+
+    ngOnDestroy(): void {
+        super.ngOnDestroy();
+        this.activeG?.remove();
+        this.activeG = null;
     }
 }
